re-enable -allowAdjustCoordsAndCell option
It seems that for certain inputs the voronoi volume differs from the
domain volume by more than the hardcoded tolerance of 0.001%.

```
...
Using voro++ with radii for particles.
Performing Voronoi decomposition.
Volume check:
  Total domain volume  = 358.306039
  Total Voronoi volume = 358.035855
Error: Voronoi volume check failed (0.075% error, > 0.001% tolerance).
```

In such cases, adding the option `-allowAdjustCoordsAndCell` tries to
move the atomic coordinates by a minute amount and reruns voro++.

Using the option was broken because it was unknown to parts of the
command line interface implementation.